requirements.txt
aiohttp==3.7.4
fastapi==0.78.0
fastapi-mail==0.1.6
langdetect==1.0.9
openai==0.27.0
pandas==1.3.0
pyjwt==2.1.0
pydantic==1.8.2
requests==2.25.0
supabase==0.2.0
python-dotenv==0.19.0
tqdm==4.61.0
uuid==1.30
